backend/requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.1
bce-python-sdk==0.9.59
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.1
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==25.12.19
fonttools==4.60.0
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.10
ImageIO==2.37.2
imagesize==1.4.1
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
modelscope==1.34.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.24.3
onnxruntime==1.23.2
opencv-contrib-python==4.10.0.84   
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86  
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.3
paddlepaddle==3.3.0
paddlex==3.3.13
pandas==2.3.3
Pillow==10.0.0
polars==1.33.1
prettytable==3.17.0
protobuf==6.33.4
psutil==7.1.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
pyparsing==3.2.5
pypdfium2==5.3.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
thop==0.1.1.post2209072238
tifffile==2025.5.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.0.196
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==2.3.7